requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
python-magic-bin==0.4.14
tensorflow==2.16.1
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
scikit-learn>=1.0.0
yara-python==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
gunicorn==21.2.0
prometheus-client==0.19.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
pylint==3.0.2
dlib==19.24.2
face-recognition==1.3.0
pytz==2024.1
